requirements.txt
discord.py
requests
python-dotenv
flask
openai
gradio_client